requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6<2.9
drf-spectacular>=0.23.1,<0.24.1
djangorestframework-simplejwt>=5.2.0,<5.2.1
pillow>=8.2.0,<8.3.0
